requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
pyttsx3
sqlalchemy
python-multipart
python-dotenv
pyannote.audio
